Add theme switch button to settings
